pyproject.toml
[project]
name = "ai-employee-sentinels"
version = "0.4.0"
description = "AI Employee Vault — Gold Tier + BaseWatcher pattern, WhatsApp watcher, Finance watcher"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=4.0.0",
    "fastmcp>=2.0.0",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.130.0",
    "tweepy>=4.14.0",
    "playwright>=1.40.0",
]

[project.scripts]
watch-inbox    = "sentinels.file_watcher:main"
watch-gmail    = "sentinels.gmail_watcher:main"
watch-linkedin = "sentinels.linkedin_poster:main"
run-scheduler  = "sentinels.scheduler:main"
run-mcp-comms    = "mcp_servers.communications:main"
run-mcp-odoo     = "mcp_servers.odoo:main"
run-mcp-social   = "mcp_servers.social_media:main"
watch-odoo       = "sentinels.odoo_watcher:main"
watch-social     = "sentinels.social_media_watcher:main"
watch-whatsapp   = "sentinels.whatsapp_watcher:main"
watch-finance    = "sentinels.finance_watcher:main"
watch-drop       = "sentinels.drop_watcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sentinels", "mcp_servers"]
